Fix PiOS aarch64 cross build, enable drm-kms-egl on bookworm, add CI

+++ shell/input/drm_seat.cc
@@ -15,6 +15,7 @@
  */
 
 #include "input/drm_seat.h"
+#include "logging/logging.h"
 
 #include <fcntl.h>
 #include <linux/input-event-codes.h>
@@ -130,7 +131,7 @@ void InstallBackstop() {
   struct sigaction sa{};
   sa.sa_handler = &BackstopFatalHandler;
   sigemptyset(&sa.sa_mask);
-  sa.sa_flags = SA_RESETHAND;
+  sa.sa_flags = static_cast<int>(SA_RESETHAND);
   for (const int sig : {SIGSEGV, SIGABRT, SIGILL, SIGFPE, SIGBUS}) {
     sigaction(sig, &sa, nullptr);
   }
